engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skuldbot-engine"
version = "0.1.0"
description = "Motor de ejecución RPA compartido para Studio y BotRunner"
authors = [
    {name = "Khipus Team", email = "dev@khipus.io"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["rpa", "automation", "robot-framework", "workflow"]

dependencies = [
    "robotframework>=6.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
rpa = [
    "rpaframework>=27.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/khipus/skuldbot-engine"
Documentation = "https://skuldbot.readthedocs.io"
Repository = "https://github.com/khipus/skuldbot-engine"

[tool.setuptools]
packages = ["skuldbot"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=skuldbot --cov-report=term-missing"

